Move handover_config into ensembl_prodinf
